prevent stalling all broadcasts when single connection is blocked
Previously a GossipChannel had one GossipSender per broadcast source
peer. That sender would stall when any of the next-hop connections
were blocked, causing broadcasts from the same source peer to be
accumulate (via GossipData.Merge) locally even though it could be sent
down the remaining unblocked connections.

We can't just have a GossipSender per connection, because broadcast
messages carry the source peer and hence only messages from the same
source peer can be merged.

The obvious fix would be for GossipChannels to have a GossipSender per
combination of source peer and connection. But that is potentially
O(n_peers * n_connections) GossipSenders, which in a fully connected
mesh is O(n_peers^2).

So instead we do the following:

1. We attempt to send a broadcast message on all its next-hop
connections. We introduce a per-connection sendTCP loop for that
purpose, which enables the detection of backlogs through failed
channel writes.

2. If we were unable to send the broadcast on one of the next-hop
connections then we propogate the broadcast message as ordinary gossip
on that connection, using the GossipSender machinery we already have
in place for that.

Note that step 2 is an optimisation only. The whole gossip machinery
is designed to cope with message loss - periodic gossip will
eventually propagate any lost information. Step 2 simply makes that
happen asap instead of a potential delay of GossipInterval.

Also note that we are assuming here that GossipBroadcast and Gossip
talk about the same information. That is the case in all current uses,
and documented in the Gossiper API by reference to a "state".